requirements.txt
slack-bolt==1.18.0
slack-sdk==3.27.1
python-dotenv==1.0.1
python-docx==1.1.0
openai==1.30.1
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
flask==2.3.3
gunicorn==21.2.0
